Extract command-sending functionality from Device class to new CommandSender class to separate concerns
