+++ miio/airconditioningcompanion.py
@@ -5,7 +5,8 @@ from typing import Optional
 import click
 
 from .click_common import EnumType, command, format_output
-from .device import Device, DeviceException
+from .device import Device
+from .exceptions import DeviceException
 
 _LOGGER = logging.getLogger(__name__)
