+++ miio/aqaracamera.py
@@ -15,7 +15,8 @@ import attr
 import click
 
 from .click_common import command, format_output
-from .device import Device, DeviceException
+from .device import Device
+from .exceptions import DeviceException
 
 _LOGGER = logging.getLogger(__name__)
