+++ miio/ceil_cli.py
@@ -5,6 +5,7 @@ import click
 
 import miio  # noqa: E402
 from miio.click_common import ExceptionHandlerGroup, validate_ip, validate_token
+from miio.command_sender import CommandSender
 
 _LOGGER = logging.getLogger(__name__)
 pass_dev = click.make_pass_decorator(miio.Ceil)
@@ -65,7 +66,7 @@ def cli(ctx, ip: str, token: str, debug: int):
 @cli.command()
 def discover():
     """Search for plugs in the network."""
-    miio.Ceil.discover()
+    CommandSender.discover()
 
 
 @cli.command()
