+++ miio/philips_rwread.py
@@ -6,7 +6,8 @@ from typing import Any, Dict
 import click
 
 from .click_common import EnumType, command, format_output
-from .device import Device, DeviceException
+from .device import Device
+from .exceptions import DeviceException
 
 _LOGGER = logging.getLogger(__name__)
