+++ miio/plug_cli.py
@@ -7,6 +7,7 @@ import click
 
 import miio  # noqa: E402
 from miio.click_common import ExceptionHandlerGroup, validate_ip, validate_token
+from miio.command_sender import CommandSender
 
 _LOGGER = logging.getLogger(__name__)
 pass_dev = click.make_pass_decorator(miio.ChuangmiPlug)
@@ -45,7 +46,7 @@ def cli(ctx, ip: str, token: str, debug: int):
 @cli.command()
 def discover():
     """Search for plugs in the network."""
-    miio.ChuangmiPlug.discover()
+    CommandSender.discover()
 
 
 @cli.command()
