+++ miio/tests/test_airqualitymonitor.py
@@ -10,7 +10,7 @@ from miio.airqualitymonitor import (
     AirQualityMonitorStatus,
 )
 
-from .dummies import DummyDevice
+from .dummies import DummyCommandSender, DummyDevice
 
 
 class DummyAirQualityMonitorV1(DummyDevice, AirQualityMonitor):
@@ -27,12 +27,14 @@ class DummyAirQualityMonitorV1(DummyDevice, AirQualityMonitor):
             "night_end_time": "format unknown",
             "sensor_state": "format unknown",
         }
-        self.return_values = {
-            "get_prop": self._get_state,
-            "set_power": lambda x: self._set_state("power", x),
-            "set_time_state": lambda x: self._set_state("time_state", x),
-            "set_night_state": lambda x: self._set_state("night_state", x),
-        }
+        self.command_sender = DummyCommandSender(
+            {
+                "get_prop": self._get_state,
+                "set_power": lambda x: self._set_state("power", x),
+                "set_time_state": lambda x: self._set_state("time_state", x),
+                "set_night_state": lambda x: self._set_state("night_state", x),
+            }
+        )
         super().__init__(args, kwargs)
 
 
@@ -94,7 +96,7 @@ class DummyAirQualityMonitorS1(DummyDevice, AirQualityMonitor):
             "temperature": 27.4,
             "tvoc": 254,
         }
-        self.return_values = {"get_prop": self._get_state}
+        self.command_sender = DummyCommandSender({"get_prop": self._get_state})
         super().__init__(args, kwargs)
 
     def _get_state(self, props):
@@ -144,7 +146,7 @@ class DummyAirQualityMonitorB1(DummyDevice, AirQualityMonitor):
             "tvoc": 1.3948699235916138,
             "tvoc_unit": "mg_m3",
         }
-        self.return_values = {"get_air_data": self._get_state}
+        self.command_sender = DummyCommandSender({"get_air_data": self._get_state})
         super().__init__(args, kwargs)
 
     def _get_state(self, props):
