+++ miio/tests/test_chuangmi_ir.py
@@ -8,7 +8,7 @@ import pytest
 from miio import ChuangmiIr
 from miio.chuangmi_ir import ChuangmiIrException
 
-from .dummies import DummyDevice
+from .dummies import DummyCommandSender, DummyDevice
 
 with open(os.path.join(os.path.dirname(__file__), "test_chuangmi_ir.json")) as inp:
     test_data = json.load(inp)
@@ -17,11 +17,13 @@ with open(os.path.join(os.path.dirname(__file__), "test_chuangmi_ir.json")) as i
 class DummyChuangmiIr(DummyDevice, ChuangmiIr):
     def __init__(self, *args, **kwargs):
         self.state = {"last_ir_played": None}
-        self.return_values = {
-            "miIO.ir_learn": lambda x: True,
-            "miIO.ir_read": lambda x: True,
-            "miIO.ir_play": self._ir_play_input_validation,
-        }
+        self.command_sender = DummyCommandSender(
+            {
+                "miIO.ir_learn": lambda x: True,
+                "miIO.ir_read": lambda x: True,
+                "miIO.ir_play": self._ir_play_input_validation,
+            }
+        )
         super().__init__(args, kwargs)
 
     def _ir_play_input_validation(self, payload):
