+++ miio/vacuum.py
@@ -13,7 +13,8 @@ import pytz
 from appdirs import user_cache_dir
 
 from .click_common import DeviceGroup, GlobalContextObject, LiteralParamType, command
-from .device import Device, DeviceException
+from .device import Device
+from .exceptions import DeviceException
 from .vacuumcontainers import (
     CarpetModeStatus,
     CleaningDetails,
@@ -596,10 +597,12 @@ class Vacuum(Device):
         @dg.resultcallback()
         @dg.device_pass
         def cleanup(vac: Vacuum, *args, **kwargs):
-            if vac.ip is None:  # dummy Device for discovery, skip teardown
+            if (
+                vac.command_sender.ip is None
+            ):  # dummy Device for discovery, skip teardown
                 return
             id_file = kwargs["id_file"]
-            seqs = {"seq": vac.raw_id, "manual_seq": vac.manual_seqnum}
+            seqs = {"seq": vac.command_sender.raw_id, "manual_seq": vac.manual_seqnum}
             _LOGGER.debug("Writing %s to %s", seqs, id_file)
             path_obj = pathlib.Path(id_file)
             cache_dir = path_obj.parents[0]
