+++ miio/wifirepeater.py
@@ -3,7 +3,8 @@ import logging
 import click
 
 from .click_common import command, format_output
-from .device import Device, DeviceException
+from .device import Device
+from .exceptions import DeviceException
 
 _LOGGER = logging.getLogger(__name__)
